[FRONTEND] feature(AuthUserProfileView, NonAuthUserProfileView): deletar e refatorar o UserProfileView
