+++ frontend/src/components/ExperienceItem.vue
@@ -22,7 +22,7 @@ export interface ExperienceItemProps {
 }
 
 /**
- * O provedor desse inject ẽ um componente pai. No caso o componente AuthUserProfileView ou NonAuthUserProfileView
+ * O provedor desse inject ẽ um componente pai. No caso o componente UserProfileView
  */
 const hasPermissions = inject('hasPermissions', false)
